requirements.txt
fastapi
uvicorn
openai
python-dotenv
google-search-results
trafilatura
pydantic
streamlit